Improved method map compiler's selection process
When a nonterminal is mapped, a method with name equal to nonterminal's name is searched and invoked via reflection. These methods are now constrianed to be defnied on a class derived from MethodMapCompiler, i.e. methods of the MethodMapCompiler or any of its base types will be ignored. This removes possible name clashes, for instance when nonterminal named "Equals" is present in the grammar, while the Equals method inherited from System.Object also exists on the executing compiler object.