Prioritized lexeme rules
Lexeme patterns are ordered in such way that keywords are recognized before symbols. Otherwise, a keyword (e.g. ignore, is) would be recognized as a terminal symbol name. Lexeme patterns which produce a match with equal length are accepted in order of appearance in the grammar - pattern closer to the beginnig of the patterns list will win.