Added explicit start symbol to grammar
Grammar text file must contain start: <nonterminal>; instruction to explicitly specify the start symbol. Before this, the first nonterminal appearing in the grammar was implicitly taken as the start symbol, which could lead to misinterpretation. For instance, if someone rearranges rules to a more human-readable order, then an inapropriate nonterminal could bubble up to become the first one in the grammar.