Creating Parser from ParserDefinition
ParserDefinition is just a description of a parser. It is generated by the parser generator and its main purpose is to be saved to a parser definition file. Proper parser used to be loaded from the file. With latest change, Parser can be constructed directly from the ParserDefinition object, without need to generate intermediate XML document.