Introduced delegate type for symbols which can be assigned from a lambda expression
When defining a grammar, one must be careful to never call rules recursively, because building the grammar would fail with stack overflow exception. Instead of calling a rule recursively, one would in fact subdue a lambda which calls the rule. In that way, a rule can depend on itself without ever effectively making a recursive call while building the grammar.
This mechanism has been improved now by introducing a delegate type which represents a nonterminal symbol. This delegate can be converted to the base type Symbol via an implicit conversion operator.