Removed literals from delegates that transform matched rules
When a rule is matched, each symbol in its right-hand side will be associated with a value of a certain type. However, literals are known to be strings, and their values are constant even before matching. Therefore, with latest improvement, lambda expression that is associated with a production rule needs not receive literals as its arguments. That simplifies code when defining rules.