Supporting implicit nonterminal references in reflection-based grammar
When a method in a reflection grammar contains an argument with no attribute assigned, it is assumed to refer to the nonterminal symbol with the same name (e.g. int Method(int something) is implicitly referencing method int Something(...)). By naming method arguments by actual methods, one does not need to apply the From attribute.